requirements.txt
Django==2.0.5
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
django-heroku
dj-database-url==0.5.0
freeze==1.0.10
gunicorn==19.9.0
jsonschema==2.6.0
mysqlclient==1.3.12
optional-django==0.1.0
parso==0.1.0
Pillow==5.1.0
psycopg2==2.7.6
pytest-django==3.2.1
python-dateutil==2.7.3
Unipath==1.0
python-decouple==3
retrying==1.3.3
somepackage==1.2.3
SQLAlchemy==1.2.10
virtualenv==16.1.0
whitenoise==3.3.0
zope.lifecycleevent==4.2.0
lxml==4.2.4
selenium==3.141.0
urllib3==1.22
requests==2.18.4
attrs==18.1.0
beautifulsoup4==4.6.0
h5py==2.8.0
html5lib==0.9999999
httpie==0.9.9
requests-oauthlib==0.8.0